+++ src/index.js
@@ -1,5 +1,5 @@
-import UnbxdSearchCore from "../../search-JS-core/src/index";
-// import UnbxdSearchCore from "@unbxd-ui/unbxd-search-core";
+// import UnbxdSearchCore from "../../search-JS-core/src/index";
+import UnbxdSearchCore from "@unbxd-ui/unbxd-search-core";
 import styles from '../styles/index.scss';
 import delegate from "./modules/utils/delegate";
 import options from './common/options';
@@ -63,7 +63,7 @@ class UnbxdSearch extends UnbxdSearchCore {
             this.options.extraParams.viewType = viewType;
             this.viewState.lastAction = "";
         }
-        if(this.viewState.lastAction != "pagination") {
+        if(this.viewState.lastAction != "pagination" && this.viewState.lastAction != "prev_page_loaded" && this.viewState.lastAction != "next_page_loaded") {
             this.viewState.isInfiniteStarted = false;
         }
         if(!viewType) {