+++ src/modules/pagination/actions.js
@@ -1,7 +1,9 @@
-const triggerNextPage = (context,next) =>{
-    context.viewState.lastAction = "pagination";
-    context.setPageStart(next);
-    context.getResults();
+const triggerNextPage = (context,next, action) =>{
+    if(context.options.pagination.type !== "INFINITE_SCROLL") {
+        context.viewState.lastAction = "pagination";
+        context.setPageStart(next);
+    }
+    context.getResults("", true, action);
     context.options.onEvent(context,context.events.pageNext, {
         value:next
     });
@@ -27,28 +29,18 @@ function renderNewResults(action) {
             triggerNextPage(this,next);
         }
     } else if(pagination.type === "INFINITE_SCROLL") {
-        // const next = start+rows;
-        // if(isNext){
-        //     this.viewState.isInfiniteStarted = true;
-        //     triggerNextPage(this,next);
-        // }
         if(action === this.actions.next){
             const next = start+rows;
-            
             if(isNext){
                 this.viewState.isInfiniteStarted = true;
-                triggerNextPage(this,next);
+                triggerNextPage(this,next, action);
             }
         }
         if(action === this.actions.prev){
             const prev = start-rows;
             if(isPrev){
                 this.viewState.isInfiniteStarted = true;
-                // this.viewState.lastAction = "pagination";
-                triggerNextPage(this,prev);
-                onEvent(this,this.events.pagePrev, {
-                    value:prev
-                });
+                triggerNextPage(this,prev, action);
             }
         }
     } else {